Sprint 7 Fixes
